+++ src/Client/ClientCore/Rendering/Scenes/Game/Gui/WorldEditor/WorldEditorGui.cs
@@ -18,8 +18,10 @@ using System.Numerics;
 using ImGuiNET;
 using Sovereign.ClientCore.Rendering.Gui;
 using Sovereign.ClientCore.Rendering.Materials;
+using Sovereign.ClientCore.Rendering.Scenes.Game.Gui.ResourceEditor;
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
 using Sovereign.ClientCore.Systems.ClientWorldEdit;
+using Sovereign.EngineCore.Events;
 
 namespace Sovereign.ClientCore.Rendering.Scenes.Game.Gui.WorldEditor;
 
@@ -28,22 +30,49 @@ namespace Sovereign.ClientCore.Rendering.Scenes.Game.Gui.WorldEditor;
 /// </summary>
 public class WorldEditorGui
 {
+    private readonly IEventSender eventSender;
     private readonly GuiExtensions guiExtensions;
 
     /// <summary>
     ///     Color used by help text.
     /// </summary>
-    private readonly Vector4 HelpTextColor = new(0.8f, 0.8f, 0.8f, 1.0f);
+    private readonly Vector4 helpTextColor = new(0.7f, 0.7f, 0.7f, 1.0f);
 
     private readonly MaterialManager materialManager;
+    private readonly MaterialSelectorPopup materialSelectorPopup;
+    private readonly ClientWorldEditController worldEditController;
     private readonly ClientWorldEditServices worldEditServices;
 
+    /// <summary>
+    ///     Backing buffer for modifier input field.
+    /// </summary>
+    private int modifierBuffer;
+
+    /// <summary>
+    ///     Change flag for modifier.
+    /// </summary>
+    private bool modifierChangeInProgress;
+
+    /// <summary>
+    ///     Backing buffer for Z offset input field.
+    /// </summary>
+    private int zOffsetBuffer;
+
+    /// <summary>
+    ///     Change flag for Z offset.
+    /// </summary>
+    private bool zOffsetChangeInProgress;
+
     public WorldEditorGui(ClientWorldEditServices worldEditServices, MaterialManager materialManager,
-        GuiExtensions guiExtensions)
+        GuiExtensions guiExtensions, MaterialSelectorPopup materialSelectorPopup, IEventSender eventSender,
+        ClientWorldEditController worldEditController)
     {
         this.worldEditServices = worldEditServices;
         this.materialManager = materialManager;
         this.guiExtensions = guiExtensions;
+        this.materialSelectorPopup = materialSelectorPopup;
+        this.eventSender = eventSender;
+        this.worldEditController = worldEditController;
     }
 
     /// <summary>
@@ -52,7 +81,7 @@ public class WorldEditorGui
     public void Render()
     {
         ImGui.SetNextWindowSize(new Vector2(300.0f, 150.0f), ImGuiCond.Appearing);
-        if (!ImGui.Begin("World Editor")) return;
+        if (!ImGui.Begin("World Editor", ImGuiWindowFlags.NoResize)) return;
 
         RenderMaterialControl();
         RenderZOffsetControl();
@@ -66,20 +95,49 @@ public class WorldEditorGui
     /// </summary>
     private void RenderMaterialControl()
     {
+        // Sync inputs with any backend changes.
+        if (modifierBuffer == worldEditServices.MaterialModifier)
+            modifierChangeInProgress = false;
+        else if (!modifierChangeInProgress)
+            modifierBuffer = worldEditServices.MaterialModifier;
+
         if (!ImGui.BeginTable("WorldEditMaterial", 2, ImGuiTableFlags.SizingStretchProp)) return;
+        ImGui.TableSetupColumn("", ImGuiTableColumnFlags.WidthFixed);
 
         var material = materialManager.Materials[worldEditServices.Material];
         var tile = material.MaterialSubtypes[worldEditServices.MaterialModifier];
 
         ImGui.TableNextColumn();
-        guiExtensions.TileSpriteButton("#material", tile.TopFaceTileSpriteId, TileSprite.Wildcard,
+        var needToOpenSelector = guiExtensions.TileSpriteButton("#material", tile.TopFaceTileSpriteId,
+            TileSprite.Wildcard,
             TileSprite.Wildcard, TileSprite.Wildcard, TileSprite.Wildcard);
 
         ImGui.TableNextColumn();
         ImGui.Text($"{material.MaterialName} (Material {worldEditServices.Material})");
-        ImGui.Text($"Modifier {worldEditServices.MaterialModifier}");
+
+        ImGui.Text("Modifier:");
+        ImGui.SameLine();
+        ImGui.InputInt("##mod", ref modifierBuffer);
+        if (modifierBuffer != worldEditServices.MaterialModifier)
+        {
+            // New value entered by user, validate.
+            if (modifierBuffer < 0 || modifierBuffer >= material.MaterialSubtypes.Count)
+            {
+                modifierBuffer = worldEditServices.MaterialModifier;
+            }
+            else
+            {
+                modifierChangeInProgress = true;
+                worldEditController.SetSelectedMaterial(eventSender, worldEditServices.Material, modifierBuffer);
+            }
+        }
 
         ImGui.EndTable();
+
+        if (needToOpenSelector) materialSelectorPopup.Open();
+        materialSelectorPopup.Render();
+        if (materialSelectorPopup.TryGetSelection(out var newMaterialId))
+            worldEditController.SetSelectedMaterial(eventSender, newMaterialId, worldEditServices.MaterialModifier);
     }
 
     /// <summary>
@@ -87,14 +145,37 @@ public class WorldEditorGui
     /// </summary>
     private void RenderZOffsetControl()
     {
+        // Sync input buffers with backend.
+        if (zOffsetBuffer == worldEditServices.ZOffset)
+            zOffsetChangeInProgress = false;
+        else if (!zOffsetChangeInProgress)
+            zOffsetBuffer = worldEditServices.ZOffset;
+
         ImGui.Separator();
-        ImGui.Text($"Z Offset: {worldEditServices.ZOffset}");
+        ImGui.Text("Z Offset:");
+        ImGui.SameLine();
+        ImGui.InputInt("##zoff", ref zOffsetBuffer);
+
+        // Validate and update state if needed.
+        if (zOffsetBuffer != worldEditServices.ZOffset)
+        {
+            if (zOffsetBuffer < ClientWorldEditConstants.MinZOffset ||
+                zOffsetBuffer > ClientWorldEditConstants.MaxZOffset)
+            {
+                zOffsetBuffer = worldEditServices.ZOffset;
+            }
+            else
+            {
+                zOffsetChangeInProgress = true;
+                worldEditController.SetZOffset(eventSender, zOffsetBuffer);
+            }
+        }
     }
 
     private void RenderHelp()
     {
         ImGui.Separator();
-        ImGui.TextColored(HelpTextColor, "Scroll to change material.");
-        ImGui.TextColored(HelpTextColor, "Ctrl+Scroll to change Z offset.");
+        ImGui.TextColored(helpTextColor, "Scroll to change material.");
+        ImGui.TextColored(helpTextColor, "Ctrl+Scroll to change Z offset.");
     }
 }