+++ src/Common/EngineCore/Events/Details/MaterialPairEventDetails.cs
@@ -1,5 +1,5 @@
 // Sovereign Engine
-// Copyright (c) 2023 opticfluorine
+// Copyright (c) 2024 opticfluorine
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,36 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-using MessagePack;
+using Sovereign.EngineCore.Components.Types;
 
-namespace Sovereign.EngineCore.Components.Types;
+namespace Sovereign.EngineCore.Events.Details;
 
 /// <summary>
-///     Represents a material identifier as a pair of material ID and
-///     material modifier.
+///     Event details containing a material ID and material modifier.
 /// </summary>
-[MessagePackObject]
-public class MaterialPair
+public class MaterialPairEventDetails : IEventDetails
 {
-    public MaterialPair()
-    {
-    }
-
-    public MaterialPair(int materialId, int materialModifier)
-    {
-        MaterialId = materialId;
-        MaterialModifier = materialModifier;
-    }
-
-    /// <summary>
-    ///     Material ID.
-    /// </summary>
-    [Key(0)]
-    public int MaterialId { get; set; }
-
-    /// <summary>
-    ///     Material modifier.
-    /// </summary>
-    [Key(1)]
-    public int MaterialModifier { get; set; }
+    public MaterialPair MaterialPair = new();
 }