+++ src/Common/EngineCore/Events/EventIds.cs
@@ -367,6 +367,22 @@ public enum EventId
 
     #endregion Client_State
 
+    #region Client_WorldEdit
+
+    /// <summary>
+    ///     Event sent to update the selected material for world editing.
+    /// </summary>
+    /// Associated details: MaterialPairEventDetails
+    Client_WorldEdit_SetMaterial = 10500,
+
+    /// <summary>
+    ///     Event sent to update the Z-offset for world editing.
+    /// </summary>
+    /// Associated details: GenericEventDetails(int)
+    Client_WorldEdit_SetZOffset = 10501,
+
+    #endregion
+
     #endregion Client
 
     #region Server